# Guarantee that incoming and outgoing request are measured by the same stat