<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

Apart from making the concurrency measurement code a bit more bulletproof in that there is now no way that the underlying stat can get skewed by request events being measured by different stats, this also slightly improves the overall performance since there is one less read-lock to obtain.

## Benchmark

```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkConcurrencyReporterHandler/sequential-16     266           218           -18.05%
BenchmarkConcurrencyReporterHandler/parallel-16       87.2          53.4          -38.76%

benchmark                                             old allocs     new allocs     delta
BenchmarkConcurrencyReporterHandler/sequential-16     0              0              +0.00%
BenchmarkConcurrencyReporterHandler/parallel-16       0              0              +0.00%

benchmark                                             old bytes     new bytes     delta
BenchmarkConcurrencyReporterHandler/sequential-16     0             0             +0.00%
BenchmarkConcurrencyReporterHandler/parallel-16       0             0             +0.00%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```